+++ homework_ballistics/include/MissionFactory.hpp
@@ -7,6 +7,8 @@
 #include "interfaces/ITargetsProvider.hpp"
 #include "interfaces/IConfigLoader.hpp"
 
+class TestFilesRepository;
+
 enum class SolverType { ANALYTICAL, TABLE };
 
 enum class ProviderType {
@@ -27,7 +29,11 @@ class MissionFactory {
   public:
     static std::unique_ptr<ITargetsProvider> createTargetsProvider(ProviderType providerType);
     static std::unique_ptr<IBallisticsSolver> createBallisticsSolver(SolverType solverType);
-    static std::unique_ptr<IConfigLoader> createConfigLoader(ConfigLoaderType configLoaderType, AmmoLoaderType ammoLoaderType);
-    static std::unique_ptr<ThreadTargetProvider> createThreadTargetsProvider(const DroneConfig &droneConfig);
-    static std::unique_ptr<ISimulationExport> createSimulationExport(ExportType exportType);
+    static std::unique_ptr<IConfigLoader> createConfigLoader(ConfigLoaderType configLoaderType,
+                                                             AmmoLoaderType ammoLoaderType,
+                                                             const std::shared_ptr<TestFilesRepository> &testFilesRepository);
+    static std::unique_ptr<ThreadTargetProvider> createThreadTargetsProvider(
+        const DroneConfig &droneConfig, const std::shared_ptr<TestFilesRepository> &testFilesRepository);
+    static std::unique_ptr<ISimulationExport> createSimulationExport(ExportType exportType,
+                                                                     const std::shared_ptr<TestFilesRepository> &testFilesRepository);
 };